requirements.txt
fastapi==0.111.1
pydantic==2.8.2
SQLAlchemy==2.0.31
psycopg2==2.9.9
pydantic-settings==2.3.4
python-dotenv==1.0.1
alembic==1.13.2
asyncpg==0.29.0
python-slugify==8.0.4
fastapi-pagination==0.12.26
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
redis==5.0.7
uvicorn==0.22.0
aiofiles==23.1.0